% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_CodonUsage.R
\name{computeCorrelationBackground}
\alias{computeCorrelationBackground}
\title{Compute the Correlation Between Mean Usage and Exonic Background}
\usage{
computeCorrelationBackground(
  mean,
  background,
  corr_method = "spearman",
  verbose = TRUE
)
}
\arguments{
\item{mean}{A \code{matrix} containing the mean usage across conditions. Can
be computed using \code{\link{computeMeanUsage}}.}

\item{background}{A \code{matrix} or table containing the frequencies in
exonic regions. Can be computed using
\code{\link{computeExonicBackground}}.}

\item{corr_method}{A correlation method accepted by \code{\link{cor}}.
Defaults to \code{"spearman"}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
Integer; correlation information between \code{mean} and
\code{background}.
}
\description{
This function calculates the \strong{correlation} between observed
\strong{mean usage} and the \strong{exonic background}. It provides a metric
for evaluating how much usage is driven by underlying sequence composition
versus condition-specific expression.
}
\examples{
data("default_tTEscanR_mRNA_data", package = "tTEscanR")
data("default_tTEscanR_metadata", package = "tTEscanR")

tTEscanR_obj <- createObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA"
)
tTEscanR_obj <- computeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = FALSE, reduce = 1000
)
codon_usage <- SummarizedExperiment::assay(tTEscanR_obj, "CodonUsage")
exonic_background <- computeExonicBackground(data = codon_usage)
# Input: expression count matrix, need to provide metadata & batch parameters
mean_codon_usage <- computeMeanUsage(
    data = codon_usage,
    mode = "raw",
    metadata = default_tTEscanR_metadata,
    batch = "tissue"
)
corr_back <- computeCorrelationBackground(
    mean = mean_codon_usage,
    background = exonic_background
)
}
