% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionAnalysis.R
\name{computeDEResults}
\alias{computeDEResults}
\title{Compute the DESeq2 Analysis}
\usage{
computeDEResults(
  list_data,
  metadata,
  target = NULL,
  batch = NULL,
  reference = NULL,
  reduce = 100,
  padj_threshold = 0.05,
  verbose = TRUE,
  compute_pairwise = TRUE
)
}
\arguments{
\item{list_data}{A \code{list} of matrices with features (i.e. genes, codons,
anticodons or amino acids) as rows and samples or conditions as columns.
The list can be named or unnamed.}

\item{metadata}{A \code{data.frame} with the data associated to the
conditions in the matrices of \code{list_data}. There has to be one
column with the same labels as the column names.}

\item{target}{Optional; a factor based on \code{metadata} columns to
define the comparisons to perform in a targeted analysis. Defaults to
\code{NULL}.}

\item{batch}{Optional; name of the categorical variable in \code{metadata}
to correct the data. Required if \code{dim_reduct} specified.}

\item{reference}{Optional; factor from the \code{batch} variable to use as
reference for the corrections. If not specified, the 1st factor that
appears will be used instead.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression
values that exceed R's handling capacity. Defaults to 100.}

\item{padj_threshold}{Numeric; p-value threshold used for highlighting
significant features in the volcano plot. Defaults to 0.05.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}

\item{compute_pairwise}{Logical; if \code{TRUE}, computes all the pairwise
comparisons based on the conditions included in the input data.}
}
\value{
A DESeq2 object with the normalized and vst counts.
}
\description{
Compute the DESeq2 Analysis
}
\examples{
data(default_tTEscanR_tRNA_data, default_tTEscanR_metadata)
DE_analysis <- computeDEResults(
    list_data = list(tRNA = default_tTEscanR_tRNA_data),
    metadata = default_tTEscanR_metadata, batch = "tissue"
)
}
