% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionAnalysis.R
\name{ComputeSizeCorrection}
\alias{ComputeSizeCorrection}
\title{Compute Size Correction to Account for Sequencing Depth}
\usage{
ComputeSizeCorrection(
  data,
  metadata = NULL,
  corr_factor = NULL,
  reduce = 100,
  verbose = TRUE
)
}
\arguments{
\item{data}{A \code{matrix} with features (i.e. genes, codons, anticodons or amino acids) as rows and samples or conditions as columns.}

\item{metadata}{A \code{data.frame} with the meta-information related with the conditions in \code{data}. There has to be one column with the same labels as the column names.}

\item{corr_factor}{A factor based on \code{metadata} columns to define the variable to correct for. Required if \code{runDESeq2} is \code{TRUE}.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression values that exceed R's handling capacity. Defaults to 100.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
A size corrected \code{data} matrix.
}
\description{
This function normalizes the data to correct for differences in sequencing depth across samples, ensuring comparability.
It is particularly useful prior to comparative analysis across conditions where bias may be introduced.
}
\examples{
data(subset_mRNA_data, metadata)
size_correction <- ComputeSizeCorrection(data = subset_mRNA_data, metadata = metadata,
                                         corr_factor = "tissue")
}
