% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_tTEscanR_metadata}
\alias{default_tTEscanR_metadata}
\title{Metadata of mRNA and tRNA data
This dataset contains the extra information to indicate the conditions of the data}
\usage{
data(default_tTEscanR_metadata)
}
\description{
Metadata of mRNA and tRNA data
This dataset contains the extra information to indicate the conditions of the data
}
