% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_tTEscanR_mRNA_data}
\alias{default_tTEscanR_mRNA_data}
\title{mRNA Expression Data Subset
This dataset contains an example of mRNA gene expression data with protein-coding genes as rows and cell types as columns.}
\usage{
data(default_tTEscanR_mRNA_data)
}
\description{
mRNA Expression Data Subset
This dataset contains an example of mRNA gene expression data with protein-coding genes as rows and cell types as columns.
}
