% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_tTEscanR_tRNA_data}
\alias{default_tTEscanR_tRNA_data}
\title{tRNA Expression Data Subset}
\format{
A matrix or data frame where rows represent tRNA gene names
and columns represent individual samples or cell-types.
}
\source{
Subset of preprocessed count table generated by Gao et al. (2022),
derived from raw sequencing data originally published by Domcke et
al. (2020).
}
\usage{
data(default_tTEscanR_tRNA_data)
}
\description{
A subset of preprocessed tRNA gene abindace count data used in package
vignettes and examples.
}
\references{
Gao W, Gallardo-Dodd CJ, Kutter C (2022). "Cell type-specific analysis by
single-cell profiling identifies a stable mammalian tRNA-mRNA interface and
increased translation efficiency in neurons." \emph{Genome Research},
32(1), 97–110. \doi{10.1101/gr.275944.121}

Domcke S, Hill AJ, Daza RM, et al. (2020). "A human cell atlas of fetal
chromatin accessibility." \emph{Science}, 370(6518), eaba7612.
\doi{10.1126/science.aba7612}
}
\keyword{datasets}
