% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_CodonFreq.R
\name{ExtractCodonComposition}
\alias{ExtractCodonComposition}
\title{Extract Codon Composition of Sequences}
\usage{
ExtractCodonComposition(sequences, verbose = TRUE)
}
\arguments{
\item{sequences}{A \code{list} of nucleotide sequences (character strings) from which to extract the codon composition.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
Codon frequency per gene table of the \code{sequences}.
}
\description{
This function analyzes a given set of nucleotide sequences and computes the count of each codon present.
}
\examples{
codon_composition <- ExtractCodonComposition(sequences = list("ATGCGTACG", "TTAAGGCCG"))
}
