% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_GeneralTranslation.R
\name{FeaturesToAA}
\alias{FeaturesToAA}
\title{Translates Genes Between Annotations or Codons, Anticodons and Amino Acids.}
\usage{
FeaturesToAA(
  data_to_translate,
  position = NULL,
  notation.from,
  notation.to,
  verbose = TRUE
)
}
\arguments{
\item{data_to_translate}{A \code{character} vector or \code{data.frame} containing gene names or features to be translated.}

\item{position}{Optional; either \code{"row"}, \code{"column"} or \code{<column_name>} to specify the location of the genes or features in \code{data_to_translate}. Required if \code{data_to_transalte} is a \code{data.frame}.}

\item{notation.from}{Either \code{"codon"} or \code{"anticodon"} to select the input format of the features in \code{data_to_translate}.}

\item{notation.to}{Either \code{"codon"}, \code{"anticodon"} or \code{"aa"} to select the output format of the features in \code{data_to_translate}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
Translated features (codons, anticodons or amino acids) from \code{data_to_translate}.
}
\description{
This function converts: codons and anticodons into anticodons, codons or amino acids based on the genetic code.
}
\examples{
data(subset_mRNA_data, metadata)
tTEscanR_obj <- Create_tTEscanR_Object(counts = subset_mRNA_data, assay = "mRNA",
                                       meta.data = list(metadata, "tissue"),
                                       meta.data.ids = list("ConditionsLabels", "CorrectionFactor"))
tTEscanR_obj <- ComputeCodonUsage(object = tTEscanR_obj, codon_freq = NULL,
                                  species = "hg38", filter = "canonical",
                                  additional.metrics = FALSE, reduce = 1000)
codons <- rownames(tTEscanR_obj@assays$CodonUsage)
codons_to_AA <- FeaturesToAA(data_to_translate = codons, notation.from = "codon",
                             notation.to = "aa")
codons_to_anticodons <- FeaturesToAA(data_to_translate = codons, notation.from = "codon",
                                     notation.to = "anticodon")
}
