% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralProcessing.R
\name{Get_tRNAMatrix}
\alias{Get_tRNAMatrix}
\title{Generate a tRNA expression matrix}
\usage{
Get_tRNAMatrix(
  data,
  assay = "peaks",
  confidence_set,
  tRNA_name_map,
  flanking_region = 100,
  name_sep = c("-", "-"),
  save = TRUE,
  out_name = NULL,
  out_directory = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{\code{SummarizedExperiment}, \code{ChromatinAssay}, or \code{SeuratObject}.}

\item{assay}{Optional; a character string specifying the name of the assay to retrieve from \code{data} if it is a \code{SeuratObject}. Defaults to \code{"peaks"}.}

\item{confidence_set}{A directory path to tRNA annotations (confidence set) in .ss or .bed format.}

\item{tRNA_name_map}{Optional; a \code{data.frame} with the tRNA gene names linked to \code{tRNA_annotations}.}

\item{flanking_region}{Integer; number of nucleotides that form the flanking region of each tRNA. Defaults to 100.}

\item{name_sep}{A string delimiter to format the tRNA gene names in the output matrix. Defaults to \code{c("-", "-")}.}

\item{save}{Logical; if \code{TRUE} stores the generated tRNA matrix into a file.}

\item{out_name}{Optional; name for the saved plot (if \code{save} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be saved (if \code{save} specified).}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
Sparse matrix of tRNA counts (tRNAs x cells)
}
\description{
Generate a tRNA expression matrix
}
