% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanRObject.R
\name{getMetadata}
\alias{getMetadata}
\alias{getMetadata,tTEscanR_Object-method}
\title{Get Metadata from a tTEscanR Object}
\usage{
getMetadata(object, name)

\S4method{getMetadata}{tTEscanR_Object}(object, name)
}
\arguments{
\item{object}{A \code{tTEscanR_Object}.}

\item{name}{Optional; A character string specifying the name of the metadata
to retrieve (e.g. "ConditionsLabels", "CorrectionFactor").}
}
\value{
A \code{data.frame} or a vector depending of the \code{name}
parameter.
}
\description{
This function safely retrieves the specified metadata from a
\code{tTEscanR_Object}.
}
\examples{
data(default_tTEscanR_mRNA_data, default_tTEscanR_metadata)
tTEscanR_obj <- CreateObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA",
    meta.data = default_tTEscanR_metadata,
    meta.data.ids = "ConditionsLabels"
)
conditions <- getMetadata(tTEscanR_obj, "ConditionsLabels")
}
