% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_GeneralProcessing.R
\name{MergeLargeMatrices}
\alias{MergeLargeMatrices}
\title{Combine large matrices}
\usage{
MergeLargeMatrices(...)
}
\arguments{
\item{...}{Variable number of \code{matrix}.}
}
\value{
Single sparse matrix with as a combination of all the input matrices.
}
\description{
This function efficiently combines individual matrices.
}
\examples{
df1 <- matrix(c(1, 0, 0, 2), nrow = 2, dimnames = list(c("geneA", "geneB"), c("s1", "s2")))
df2 <- matrix(c(3, 0, 0, 4), nrow = 2, dimnames = list(c("geneB", "geneC"), c("s2", "s3")))
merged_matrix <- MergeLargeMatrices(df1, df2)
}
