% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata of mRNA and tRNA data
This dataset contains the extra information to indicate the conditions of the data}
\usage{
data(metadata)
}
\description{
Metadata of mRNA and tRNA data
This dataset contains the extra information to indicate the conditions of the data
}
