% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA_data}
\alias{mRNA_data}
\title{mRNA Expression Data
This dataset contains mRNA expression levels for a sample of genes (symbol name) across various cell types.}
\usage{
data(mRNA_data)
}
\description{
mRNA Expression Data
This dataset contains mRNA expression levels for a sample of genes (symbol name) across various cell types.
}
