% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonFreqPerGene.R
\name{ObtainCodonFreqPerGene}
\alias{ObtainCodonFreqPerGene}
\title{Compute Codon Frequency-per-Gene Table}
\usage{
ObtainCodonFreqPerGene(
  dataset_name = NULL,
  genes_file = NULL,
  transcripts = NULL,
  filter = "canonical",
  retain.mitochondrial = FALSE,
  out_format = "external_gene_name",
  verbose = TRUE
)
}
\arguments{
\item{dataset_name}{A character string specifying the Ensembl species dataset name (e.g. \code{"hsapiens_gene_ensembl"}).}

\item{genes_file}{Optional; a path to a FASTA file.}

\item{transcripts}{Optional; a character vector of transcripts or gene IDs to subset the analysis.}

\item{filter}{Either \code{"canonical"} (default) or \code{"length"} (longest transcript) to specify which transcript to choose if several are available for the same gene.}

\item{retain.mitochondrial}{Logical; if \code{FALSE} filters outs the mitochondrial genes. Defaults to \code{FALSE}.}

\item{out_format}{Either \code{"external_gene_name"} (default), \code{"ensembl_gene_id"} or \code{"ensembl_transcript_id"} to specify annotation to use in the output codon frequency-per-gene table.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
Codon frequency-per-gene table and a translator gene annotation table (if available).
}
\description{
This function computes codon usage frequencies for each gene based on a provided set of gene sequences.
It can optionally subset the analysis to specific transcripts and apply filtering criteria when multiple transcripts are available per gene.
Consequently, no \code{filter}parameter will be considered if parameters \code{transcripts} and \code{genes_file} are given.
}
