% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{PlotCorrelation}
\alias{PlotCorrelation}
\title{Correlation Plot: Exonic Codon Background - Mean Codon Usage}
\usage{
PlotCorrelation(
    data,
    plot = "MeanCodonUsage",
    x_axis_col,
    y_axis_col,
    condition_col,
    extra_val = NULL,
    label_col = NULL,
    color_palette = NULL,
    out_name = NULL,
    show_legend = "none",
    targeted_arg = NULL,
    save_format = NULL,
    out_directory = NULL,
    add_titles = TRUE,
    verbose = TRUE
)
}
\arguments{
\item{data}{A long format table. This format can be obtained using
\code{\link{TransformFormat}}.}

\item{plot}{Either \code{"MeanCodonUsage"} (default) or
\code{"PoolDiversity"} to indicate the \code{data} source.}

\item{x_axis_col}{Name of the categorical variable to reflect in the plot.}

\item{y_axis_col}{Name of the numerical variable to reflect in the plot.}

\item{condition_col}{Name of the categorical variable to group the data
points.}

\item{extra_val}{Optional; variable with additional information to include
in the plot (e.g. correlation value).}

\item{label_col}{Name of the categorical variable to label the data points.}

\item{color_palette}{Optional; a vector of color codes to customize plot
appearance.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format}
specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of
the legend.}

\item{targeted_arg}{Optional; a vector defining key feature clusters to
highlight or label.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify
the format to save the plot.}

\item{out_directory}{Optional; path to the directory where the plot will
be saved (if \code{save_format} specified).}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot.
Defaults to \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the correlation. If
\code{save_format} is provided, the plot will also be saved to the specified
location.
}
\description{
This function generates a visualization to correlate different parameters of
the \code{data}. The input data (\code{data}) is expected to be in long
format and contain a minimum information regarding the features, the
conditions they belong to and their usage counts. A common usage of this
kind of plot is to represent the codon frequencies correlations of the
exonic background and the mean codon usage across conditions. Based on the
selected \code{plot} parameter the user can select the most convenient layout
to display the data. It is crucial to ensure consistency between the name of
the columns in \code{data} and the parameters \code{x_axis_col},
\code{y_axis_col} and \code{condition_col}.
}
\examples{
data(default_tTEscanR_mRNA_data, default_tTEscanR_metadata)

# Define the object and compute the codon usage
tTEscanR_obj <- CreateObject(
    counts = default_tTEscanR_mRNA_data, assay = "mRNA",
    meta.data = list(default_tTEscanR_metadata, "tissue"),
    meta.data.ids = list("ConditionsLabels", "CorrectionFactor")
)
tTEscanR_obj <- ComputeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = TRUE, reduce = 1000
)

# Compute and extract the mean codon usage
additional_metrics <- getMetadata(
    tTEscanR_obj,
    "CodonUsage_AdditionalMetrics"
)
mean_codon_usage <- additional_metrics$MeanCodonUsage
exonic_background <- additional_metrics$CodonExonicBackground
exonic_background <- as.data.frame(exonic_background)
correlation_mean_background <- cbind(mean_codon_usage, exonic_background)

PlotCorrelation(
    data = correlation_mean_background, plot = "MeanCodonUsage",
    x_axis_col = "mean_usage_across_conditions",
    y_axis_col = "exonic_background", condition_col = "feature",
    extra_val = additional_metrics$MeanCodonCorr, add_titles = TRUE,
    show_legend = "none"
)
}
