% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{plotPermutation}
\alias{plotPermutation}
\title{Permutation Plot}
\usage{
plotPermutation(
  permut_data,
  sig_data,
  color_palette = NULL,
  filename = NULL,
  show_legend = "none",
  add_titles = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{permut_data}{A table with the codons and their frequencies after
computing all the permutations. Output from
\code{\link{getPermutationDist}}.}

\item{sig_data}{A table with the codon exonic background and their
significance level before (p-value) and after the correction (p-adjusted
value). Output from \code{\link{obtainSignificance}}}

\item{color_palette}{Optional; a vector of color codes to customize plot
appearance.}

\item{filename}{Optional; name for the plot to be saved.}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the
position of the legend in the plot.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot.
Defaults to \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
Permutation plot.
}
\description{
This function generates a plot to compare the baseline codon exonic
background against the current codon usage. For a better interpretation, the
codons are colored by amino acid.
}
\examples{
data("default_tTEscanR_mRNA_data", package = "tTEscanR")
data("default_tTEscanR_metadata", package = "tTEscanR")

selected_genes <- default_tTEscanR_mRNA_data[1:20, ]
permutation_test <- getPermutationDist(
    n_permut = 100, target_data = selected_genes, species = "hg38"
) # Generate table with codon and freq
tTEscanR_obj <- createObject(
    counts = default_tTEscanR_mRNA_data, assay = "mRNA",
    meta_data = default_tTEscanR_metadata,
    sample_id = "conditions",
    params = list("CorrectionFactor" = "tissue")
)
tTEscanR_obj <- computeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = FALSE, reduce = 1000
)

codon_usage <- SummarizedExperiment::assay(tTEscanR_obj, "CodonUsage")
codon_background <- rowSums(codon_usage) / sum(rowSums(codon_usage))
codons_to_AA <- featuresToAA(
    data = names(codon_background),
    notation_from = "codon", notation_to = "aa"
)
codon_background <- data.frame(
    group = codons_to_AA, codon = names(codon_background),
    freq = as.numeric(codon_background), row.names = NULL
)
significance <- obtainSignificance(
    dist = permutation_test, value = codon_background
)

plotPermutation(permut_data = permutation_test, sig_data = significance)
}
