% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{plotTEscore}
\alias{plotTEscore}
\title{Violin Plot Displaying tTE Scores}
\usage{
plotTEscore(
  data,
  metadata,
  class_col,
  index_col,
  target_col = NULL,
  score_col = "tTE",
  cond_col = "condition",
  pval_col = "p_value",
  facet_col = NULL,
  color_palette = NULL,
  save_format = NULL,
  out_name = NULL,
  add_stats = TRUE,
  out_directory = NULL,
  verbose = TRUE,
  show_legend = "none",
  add_titles = TRUE,
  show_outliers = FALSE
)
}
\arguments{
\item{data}{A tTE results table obtained from \code{\link{computeTheoreticalTE}}.}

\item{metadata}{A table with additional information regarding the conditions
in \code{data}.}

\item{class_col}{Name of the categorical variable to reflect in the plot.}

\item{index_col}{Name of the categorical variable that links the conditions
in \code{data} with the \code{metadata}.}

\item{target_col}{Optional; name of the categorical variable to perform the
statistical comparison (the most specific level). Used if
\code{add_stats} is \code{TRUE}.}

\item{score_col}{Name of the numerical variable that contains the tTE scores
in the \code{data}. Defaults to \code{"tTE"}.}

\item{cond_col}{Name of the categorical variable that contains the conditions
in the \code{data}. Defaults to \code{"condition"}.}

\item{pval_col}{Name of the numerical variable that contains the significance
scores in \code{data}. Defaults to \code{"p_value"}.}

\item{facet_col}{Optional; name of the categorical variable separate the
plot into different panels.}

\item{color_palette}{Optional; a vector of color codes to customize plot
appearance.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify
the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format}
specified).}

\item{add_stats}{Logical; if \code{TRUE}, performs a statistical analysis
based on the available parameters. Defaults to \code{TRUE}.}

\item{out_directory}{Optional; path to the directory where the plot will be
saved (if \code{save_format} specified).}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the
position of the legend in the plot.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot.
Defaults to \code{TRUE}.}

\item{show_outliers}{Logical; if \code{TRUE}, labels the outlier data points
based on \code{index_col}.}
}
\value{
A \code{ggplot} object representing the tTE scores. If
\code{save_format} is provided, the plot will also be saved to the
specified location. If \code{add_stats} reports a table with the
statitical measures summarized.
}
\description{
This function generates a violin plot to visualize the distribution of tTE
scores across different condition. The tTE scores should be obtained using
\code{\link{computeTheoreticalTE}}. The plot allows the user to easily compare the
tTE distribution between different conditions, with the option to highlight
specific feature clusters based on the provided \code{targets}.
}
\examples{
data(
    default_tTEscanR_mRNA_data, default_tTEscanR_tRNA_data,
    default_tTEscanR_metadata
)

# Define the tTEscanR object
tTEscanR_obj <- createObject(
    counts = list(
        mRNA = default_tTEscanR_mRNA_data, tRNA = default_tTEscanR_tRNA_data
    ),
    meta.data = list(default_tTEscanR_metadata, "tissue"),
    meta.data.ids = list("ConditionsLabels", "CorrectionFactor")
)

# Compute the codon and anticodon usage
tTEscanR_obj <- computeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = FALSE, reduce = 10000
)
tTEscanR_obj <- computeAnticodonUsage(object = tTEscanR_obj)

# Compute the theoretical translation efficiency (tTE scores)
tTEscanR_obj <- computeTheoreticalTE(
    object = tTEscanR_obj, level = "codon", compute_significance = TRUE
)
tTEresults_codon <- getMetadata(tTEscanR_obj, "tTEresults_codon")
conditions_metadata <- getMetadata(tTEscanR_obj, "ConditionsLabels")

# Visualize the tTE scores
plotTEscore(
    data = tTEresults_codon, metadata = conditions_metadata,
    index_col = "conditions", class_col = "tissue", add_stats = TRUE
)
}
