% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanRGeneral.R
\name{runPipeline}
\alias{runPipeline}
\title{Runs the Theoretical Translation Efficiency (tTE) Pipeline}
\usage{
runPipeline(
  mRNA_data,
  tRNA_data,
  meta_data,
  batch,
  sample_id = NULL,
  corr_method = "spearman",
  additional_metrics = TRUE,
  runDESeq = TRUE,
  compute_significance = TRUE,
  codon_freq = NULL,
  species = NULL,
  genetic_code = "Standard",
  reduce = 100,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{mRNA_data}{A count matrix of mRNA genes (rows) per conditions
(columns).}

\item{tRNA_data}{A count matrix of tRNA genes (rows) per conditions
(columns).}

\item{meta_data}{A \code{data.frame} with the meta-information related with
the conditions in \code{mRNA_data} or \code{tRNA_data}. Just the
intersecting conditions will be considered.}

\item{batch}{A factor based on \code{meta_data} columns to define the
variable to correct for when size correcting the count matrices.}

\item{sample_id}{Optional; a character string naming the column in
\code{meta.data} to use as sample row names.}

\item{corr_method}{A correlation method accepted by \code{\link{cor}}.
Defaults to \code{"spearman"}.}

\item{additional_metrics}{Logical; if \code{TRUE}, computes: (i) codon
exonic background, (ii) mean codon usage, and (iii) correlation
between the previous metrics. Defaults to \code{TRUE}.}

\item{runDESeq}{Logical; if \code{TRUE}, performs differential expression
analysis to each assay in the \code{tTEscanR_Object}. Defaults to
\code{TRUE}}

\item{compute_significance}{Logical; if \code{TRUE}, computes the statistical
significance (p-value) of the tTE scores. Defaults to \code{TRUE}.}

\item{codon_freq}{Optional; a user-provided codon frequency-per-gene table.
If necessary, it can be computed using \code{\link{getCodonFreq}}.}

\item{species}{Either \code{"hg38"} (human) or \code{"mm39"} (mouse) to
specify which default codon frequency-per-gene table to use. Required if
\code{codon_freq} not provided or if the gene annotation is inconsistent
between both inputs.}

\item{genetic_code}{A \code{character} string to specify the genetic code
to be used. Defaults to \code{"Standard"}.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression
values that exceed R's handling capacity. Defaults to 100.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed directly to
\code{\link{runDEAnalysis}} (e.g., \code{dim_reduct},
\code{compute_pairwise}, \code{color_factor}).}
}
\value{
A \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} with the
assays and metadata computed through the tTE pipeline.
}
\description{
This function wraps up all the independent functions of theoretical
translation efficiency pipeline. Requires an mRNA and tRNA count matrices to
compute the codon and anticodon usage and further derive the amino acid
demand and supply. With matching condition in the former matrices the
theoretical translation efficiency would be computed.
}
\examples{
data("default_tTEscanR_mRNA_data", package = "tTEscanR")
data("default_tTEscanR_tRNA_data", package = "tTEscanR")
data("default_tTEscanR_metadata", package = "tTEscanR")

tTEscanR_obj <- runPipeline(
    mRNA_data = default_tTEscanR_mRNA_data,
    tRNA_data = default_tTEscanR_tRNA_data,
    meta_data = default_tTEscanR_metadata,
    species = "hg38", batch = "tissue", additional_metrics = FALSE,
    sample_id = "conditions", compute_significance = FALSE, runDESeq = FALSE
)
}
