% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralProcessing.R
\name{Set_tRNACutoff}
\alias{Set_tRNACutoff}
\title{Selection of the Optimal tRNA Cut Cutoff}
\usage{
Set_tRNACutoff(
  data,
  num_iter = 1000,
  cutoffs_limits = c(50, 10000),
  generate_plot = TRUE,
  slope_threshold = 0.001,
  rho_threshold = 0.95,
  compute_aa = FALSE
)
}
\arguments{
\item{data}{tRNA gene expression count \code{matrix} with tRNA genes as rows and conditions as columns.}

\item{num_iter}{Numeric; value to select the number of iterations to perform in order to determine the optimal cutoff. Defaults to 1000.}

\item{cutoffs_limits}{Minimum and maximum values to test to search for the optimal tRNA cuts threshold. Defaults to c(50, 10000).}

\item{generate_plot}{Logic; if \code{TRUE}, generates a correlation plot. Defaults to \code{TRUE}.}

\item{slope_threshold}{Numeric; value to consider for the determination of the correlation stability. Defaults to 0.001.}

\item{rho_threshold}{Numeric; value to consider for the determination of the correlation strength. Defaults to 0.95.}

\item{compute_aa}{Logic; if \code{TRUE}, computes the amino acid supply, otherwise only considers the anticodon usage. Defaults to \code{FALSE}.}
}
\value{
Table with the optimal cutoff at the anticodon isoacceptor and amino acid isotype.
}
\description{
Selection of the Optimal tRNA Cut Cutoff
}
\examples{
data(default_tTEscanR_tRNA_data)
optimal_tRNA_cutoffs <- Set_tRNACutoff(data = default_tTEscanR_tRNA_data,
                                       generate_plot = FALSE, num_iter = 100,
                                       cutoffs_limits = c(3000, 5000))
}
