% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subset_mRNA_data}
\alias{subset_mRNA_data}
\title{mRNA Expression Data Subset
This dataset contains a subset of mRNA_data.}
\usage{
data(subset_mRNA_data)
}
\description{
mRNA Expression Data Subset
This dataset contains a subset of mRNA_data.
}
