% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tRNA_data}
\alias{tRNA_data}
\title{tRNA Expression Data
This dataset contains tRNA expression levels for a sample of genes across various cell types.}
\usage{
data(tRNA_data)
}
\description{
tRNA Expression Data
This dataset contains tRNA expression levels for a sample of genes across various cell types.
}
