% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralProcessing.R
\name{tRNACutsCutoff}
\alias{tRNACutsCutoff}
\title{Selection of the Optimal tRNA Cut Cutoff}
\usage{
tRNACutsCutoff(
  data,
  cutoffs = seq(50, 10000, by = 50),
  generate_plot = TRUE,
  slope_threshold = 0.001
)
}
\arguments{
\item{data}{tRNA gene expression count \code{matrix} with tRNA genes as rows and conditions as columns.}

\item{cutoffs}{Set of values to test to search for the optimal tRNA cuts threshold. Defaults to seq(50, 10000, by = 50).}

\item{generate_plot}{Logic; if \code{TRUE}, generates a correlation plot. Defaults to \code{TRUE}.}

\item{slope_threshold}{Numeric; value to consider for the determination of the correlation stability.}
}
\value{
Table with the optimal cutoff at the anticodon isoacceptor and amino acid isotype.
}
\description{
Selection of the Optimal tRNA Cut Cutoff
}
\examples{
data(subset_tRNA_data)
optimal_tRNA_cutoffs <- tRNACutsCutoff(data = subset_tRNA_data,
                                       cutoffs = seq(50, 100, by = 10),
                                       generate_plot = FALSE)
}
