% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralProcessing.R
\name{tRNASetGenes}
\alias{tRNASetGenes}
\title{Annotate the tRNA genes from tRNA tags}
\usage{
tRNASetGenes(data, tRNA_bed, flanking_region = 100, name_sep = c("-", "-"))
}
\arguments{
\item{data}{A \code{matrix} with tRNA tags as rows.}

\item{tRNA_bed}{Path to the directory that contains the .bed file}

\item{flanking_region}{Numeric; number of bases to include expand the region
interrogated. Defaults to 100.}

\item{name_sep}{A string delimiter to format the tRNA gene names in the
output matrix. Defaults to \code{c("-", "-")}.}
}
\value{
A \code{data} with the translated tRNA gene names.
}
\description{
Annotate the tRNA genes from tRNA tags
}
