% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{tTE_CompareTargetToMean}
\alias{tTE_CompareTargetToMean}
\title{Distribution Plot of Codon/Anticodon Usage or Amino Acid Demand/Supply Compared to the Mean}
\usage{
tTE_CompareTargetToMean(
  target_data,
  overall_data,
  x_axis_col,
  y_axis_col,
  show_difference = TRUE,
  color_palette = NULL,
  save_format = NULL,
  out_name = NULL,
  out_directory = NULL,
  show_legend = "none",
  add_titles = TRUE
)
}
\arguments{
\item{target_data}{A long format table of a condition of interest. This format can be obtained using \code{\link{DataToLongFormat}}.}

\item{overall_data}{A long format table of the whole set of conditions (with or without the data in \code{target_data}. This format can be obtained using \code{\link{DataToLongFormat}}.}

\item{x_axis_col}{Name of the categorical variable to reflect in the plot.}

\item{y_axis_col}{Name of the numerical variable to reflect in the plot.}

\item{show_difference}{Logical; if \code{TRUE}, displays the difference between \code{overall_data} and \code{target_data}.}

\item{color_palette}{Optional; a vector of color codes (min. 2 codes and max. 3 codes) to customize plot appearance. Colors for (i) data, (ii) target value, and (iii) difference bar.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"}, \code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of the legend.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the requested distribution plot. If \code{save_format} is provided, the plot will also be saved to the specified location.
}
\description{
This function generates a distribution plot to compare distribution usages between a targeted set of conditions against the rest.
A common application of this plot is to compare mean usages across conditions.
Both input data sources (\code{target_data} and \code{overall_data}) are expected to be in long format and contain a minimum information regarding the features, the conditions they belong to and their usage counts.
It is crucial to ensure consistency between the name of the columns in \code{target_data} and \code{overall_data}, together with the parameters \code{x_axis_col} and \code{y_axis_col}.
}
