% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{tTE_CorrelationPlot}
\alias{tTE_CorrelationPlot}
\title{Correlation Plot: Exonic Codon Background - Mean Codon Usage}
\usage{
tTE_CorrelationPlot(
  data,
  plot = "MeanCodonUsage",
  x_axis_col,
  y_axis_col,
  condition_col,
  label_col = NULL,
  extra_val = NULL,
  color_palette = NULL,
  targeted_arg = NULL,
  save_format = NULL,
  out_name = NULL,
  out_directory = NULL,
  show_legend = "none",
  add_titles = TRUE
)
}
\arguments{
\item{data}{A long format table. This format can be obtained using \code{\link{DataToLongFormat}}.}

\item{plot}{Either \code{"MeanCodonUsage"} (default) or \code{"PoolDiversity"} to indicate the \code{data} source.}

\item{x_axis_col}{Name of the categorical variable to reflect in the plot.}

\item{y_axis_col}{Name of the numerical variable to reflect in the plot.}

\item{condition_col}{Name of the categorical variable to group the data points.}

\item{label_col}{Name of the categorical variable to label the data points.}

\item{extra_val}{Optional; variable with additional information to include in the plot (e.g. correlation value).}

\item{color_palette}{Optional; a vector of color codes to customize plot appearance.}

\item{targeted_arg}{Optional; a vector defining key feature clusters to highlight or label.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"}, \code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of the legend.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the correlation. If \code{save_format} is provided, the plot will also be saved to the specified location.
}
\description{
This function generates a visualization to correlate different parmaeters of the \code{data}.
The input data (\code{data}) is expected to be in long format and contain a minimum information regarding the features, the conditions they belong to and their usage counts.
A common usage of this kind of plot is to represent the codon frequencies correlations of the exonic background and the mean codon usage across conditions.
Based on the selected \code{plot} parameter the user can select the most convenient layout to display the data.
It is crucial to ensure consistency between the name of the columns in \code{data} and the parameters \code{x_axis_col}, \code{y_axis_col} and \code{condition_col}.
}
