% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{tTE_PermutationPlot}
\alias{tTE_PermutationPlot}
\title{Permutation Plot}
\usage{
tTE_PermutationPlot(
  permut_data,
  sig_data,
  color_palette = NULL,
  save_format = NULL,
  out_name = NULL,
  out_directory = NULL,
  show_legend = "none",
  add_titles = TRUE
)
}
\arguments{
\item{permut_data}{A table with the codons and their frequencies after computing all the permutations. Output from \code{\link{GetPermutationDist}}.}

\item{sig_data}{A table with the codon exonic background and their significance level before (p-value) and after the correction (p-adjusted value). Output from \code{\link{ObtainSignificance}}}

\item{color_palette}{Optional; a vector of color codes to customize plot appearance.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"}, \code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of the legend in the plot.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot. Defaults to \code{TRUE}.}
}
\value{
Permutation plot.
}
\description{
This function generates a plot to compare the baseline codon exonic background against the current codon usage.
For a better interpretation, the codons are colored by amino acid.
}
