% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanR-package.R
\docType{package}
\name{tTEscanR-package}
\alias{tTEscanR}
\alias{tTEscanR-package}
\title{tTEscanR: A user-friendly integrative R package for quantifying and visualizing translation efficiency from sequencing data in diverse biological systems}
\description{
Translation elongation is dependent on codon-anticodon interactions, with suitable nucleotide pairing being essential for efficient translation. To quantify this relationship, we previously developed a computational pipeline (GitHub - wgao688/sc_tRNA_mRNA) that uses mRNA codon usage relative to tRNA anticodon availability as a proxy for theoretical translation efficiency (tTE). Here, we introduce tTEscanR, a powerful and user-friendly R-based package that extends this approach to quantify translation efficiency from both bulk and single-cell sequencing data. tTEscanR is a versatile tool for exploring translation efficiency in diverse cellular processes, disease mechanisms, and therapeutic development. It also features an advanced visualization module to generate high-quality plots, enhancing result interpretation and communication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/avarassanchez/tTEscanR}
  \item Report bugs at \url{https://github.com/avarassanchez/tTEscanR/issues}
}

}
\author{
\strong{Maintainer}: Ana Varas-Sánchez \email{ana.varas@scilifelab.se} (\href{https://orcid.org/0009-0006-0187-8968}{ORCID})

Authors:
\itemize{
  \item Ana Varas-Sánchez \email{ana.varas@scilifelab.se} (\href{https://orcid.org/0009-0006-0187-8968}{ORCID})
  \item Carlos Gallardo-Dodd \email{carlos.gallardo@scilifelab.se} (\href{https://orcid.org/0000-0002-6086-0550}{ORCID})
  \item William Gao \email{william.gao@pennmedicine.upenn.edu} (\href{https://orcid.org/0000-0002-9415-331X}{ORCID})
}

Other contributors:
\itemize{
  \item Claudia Kutter \email{claudia.kutter@ki.se} (\href{https://orcid.org/0000-0002-8047-0058}{ORCID}) [thesis advisor]
  \item Karolinska Institutet (grant_id: KID-2021-00308, KID-2023-01298) [funder]
  \item Swedish Research Council (grant_id: 2023-02780) [funder]
  \item Cancerfonden (grant_id: 22 2246 Pj, 25 4703 Pj) [funder]
}

}
\keyword{internal}
