% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_tTEscanRObject.R
\name{Update_tTEscanR_Object}
\alias{Update_tTEscanR_Object}
\title{tTEscanR Object Update}
\usage{
Update_tTEscanR_Object(
  object,
  counts = NULL,
  assay = NULL,
  main_name = NULL,
  meta.data = NULL,
  meta.data.ids = NULL,
  overwrite.assay = FALSE,
  overwrite.metadata = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{An existing \code{tTEscanR_Object}.}

\item{counts}{Optional; a count matrix (or \code{list} of matrices) that will be stored in the \code{assays} slot of the input \code{object}. Supported formats: \code{matrix}, \code{data.frame} and \code{list}.}

\item{assay}{Optional; a character string (or vector of strings) specifying the name of the \code{counts}. Supported formats: \code{character} and \code{list}.}

\item{main_name}{Optional; a character string specifying the name of the \code{meta.data} if dealing with a \code{list} that needs to be added as a single element.}

\item{meta.data}{Optional; a \code{list} with additional information that will be stored in \code{meta.data} slot of the input \code{object}.}

\item{meta.data.ids}{Optional; a \code{list} with the labels to identify the \code{meta.data} (if \code{meta.data} is given).}

\item{overwrite.assay}{Logical; if \code{TRUE}, overwrites any existing assay in the \code{object} if the \code{assay} label coincides. Defaults to \code{FALSE}.}

\item{overwrite.metadata}{Logical; if \code{TRUE}, overwrites any existing metadata in the \code{object} if the \code{meta.data.ids} label coincides. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
An updated \code{tTEscanR_Object}.
}
\description{
Updates an already created \code{tTEscanR_object} using \code{\link{Create_tTEscanR_Object}}. For more details, refer to \code{\link{Create_tTEscanR_Object}}.
}
\examples{
data(subset_mRNA_data, subset_tRNA_data)
tTEscanR_obj <- Create_tTEscanR_Object(counts = subset_mRNA_data, assay = "mRNA")
tTEscanR_obj <- Update_tTEscanR_Object(object = tTEscanR_obj, counts = subset_tRNA_data,
                                       assay = "tRNA")
}
