% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanRObject.R
\name{UpdateObject}
\alias{UpdateObject}
\title{tTEscanR Object Update}
\usage{
UpdateObject(
    object,
    counts = NULL,
    assay = NULL,
    main_name = NULL,
    meta.data = NULL,
    meta.data.ids = NULL,
    overwrite = FALSE,
    verbose = TRUE
)
}
\arguments{
\item{object}{An existing \code{tTEscanR_Object}.}

\item{counts}{Optional; a count matrix (or \code{list} of matrices) that
will be stored in the \code{assays} slot of the input \code{object}.
Supported formats: \code{matrix}, \code{data.frame} and \code{list}.}

\item{assay}{Optional; a \code{character} string (or \code{list} of strings)
specifying the name of the \code{counts}. Supported formats: \code{character}
and \code{list}.}

\item{main_name}{Optional; a \code{character} string specifying the name of
the \code{meta.data} if dealing with a \code{list} that needs to be added as
a single element.}

\item{meta.data}{Optional; a \code{list} with additional information that
will be stored in \code{meta.data} slot of the input \code{object}.}

\item{meta.data.ids}{Optional; a \code{list} with the labels to identify the
\code{meta.data} (if \code{meta.data} is given).}

\item{overwrite}{Logical; if \code{TRUE}, overwrites any existing assay or
metadata in the \code{object} if the \code{assay} or \code{meta.data.ids}
label coincides. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
An updated \code{tTEscanR_Object}.
}
\description{
Updates an existing \code{tTEscanR_object} using
\code{\link{CreateObject}}. For more details, refer to
\code{\link{CreateObject}}.
}
\examples{
data(default_tTEscanR_mRNA_data, default_tTEscanR_tRNA_data)
tTEscanR_obj <- CreateObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA"
)
tTEscanR_obj <- UpdateObject(
    object = tTEscanR_obj,
    counts = default_tTEscanR_tRNA_data,
    assay = "tRNA"
)
}
