% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TheoreticalTranslationEfficiency.R
\name{Compute_tTE}
\alias{Compute_tTE}
\title{Compute the Theoretical Translation Efficiency (tTE) score}
\usage{
Compute_tTE(
  object,
  level = "aa",
  corr_method = "spearman",
  compute.significance = TRUE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{tTEscanR_Object} containing a mRNA and a codon usage assay and/or a tRNA and an anticodon usage assay.}

\item{level}{Either \code{"codon"}, \code{"aa"} (default) or \code{"both"} to indicate which analysis to perform. Defaults to \code{"aa"}.}

\item{corr_method}{A correlation method accepted by \code{\link{cor}}. Defaults to \code{"spearman"}.}

\item{compute.significance}{Logical; if \code{TRUE}, computes the statistical significance (p-value) of the tTE scores. Defaults to \code{TRUE}.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites any existing tTE table in the \code{tTEscanR_Object}. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
An updated \code{tTEscanR_Object} containing a new layer of information representing the translation efficiency table for the matching conditions in the mRNA and tRNA data.
}
\description{
This function calculates the theoretical translation efficiency (tTE) score by integrating codon-anticodon usage and/or amino acid demand-supply across conditions.
}
\examples{
data(subset_mRNA_data, subset_tRNA_data, metadata)
tTEscanR_obj <- Create_tTEscanR_Object(counts = list(mRNA = subset_mRNA_data,
                                                     tRNA = subset_tRNA_data),
                                       meta.data = list(metadata, "tissue"),
                                       meta.data.ids = list("ConditionsLabels", "CorrectionFactor"))
tTEscanR_obj <- ComputeCodonUsage(object = tTEscanR_obj, species = "hg38",
                                  additional.metrics = FALSE, reduce = 10000)
tTEscanR_obj <- ComputeAnticodonUsage(object = tTEscanR_obj)
tTEscanR_obj <- Compute_tTE(object = tTEscanR_obj, level = "codon", compute.significance = FALSE)
}
