% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAUsage.R
\name{ComputeAAUsage}
\alias{ComputeAAUsage}
\title{Compute Amino Acid (AA) Demand or Supply}
\usage{
ComputeAAUsage(object, level, overwrite.assay = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A \code{tTEscanR_Object} containing codon and/or anticodon usage assays to be analyzed.}

\item{level}{Either \code{"demand"}, \code{"supply"} or \code{"both"} to indicate which analysis to perform.}

\item{overwrite.assay}{Logical; if \code{TRUE}, overwrites any existing assay in the \code{object}. Defaults to \code{FALSE}}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}}
}
\value{
An updated \code{tTEscanR_Object} containing a new layer of information in the \code{assays} slot representing the AA demand and/or supply.
}
\description{
This function calculates the amino acid (AA) demand and/or supply from a codon and/or anticodon usage matrices of a \code{tTEscanR_Object}.
It aggregates the contribution of their features based on the standard genetic code (mapping codons/anticodons to AA),
The resulting values reflect total usage (demand) or availability (supply) of each amino acid, depending on the input type.
}
\examples{
data(subset_tRNA_data)
tTEscanR_obj <- Create_tTEscanR_Object(counts = subset_tRNA_data, assay = "tRNA")
tTEscanR_obj <- ComputeAnticodonUsage(object = tTEscanR_obj)
tTEscanR_obj <- ComputeAAUsage(object = tTEscanR_obj, level = "supply")
}
