% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonUsage.R
\name{ComputeAnticodonUsage}
\alias{ComputeAnticodonUsage}
\title{Compute Anticodon Usage from tRNA Gene Expression Data}
\usage{
ComputeAnticodonUsage(object, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A \code{tTEscanR_Object} containing a tRNA assay.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites any existing assay and
metadata in the \code{object}. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
An updated \code{tTEscanR_Object} containing a new layer of
information \code{"AnticodonUsage"} in the \code{assays} slot representing
the anticodon usage.
}
\description{
This function calculates \strong{anticodon usage profiles} from tRNA gene
expression data stored in a \code{tTEscanR_Object}. It summarizes the
expression of tRNAs by their anticodon identity, which can be used to
estimate the tRNA supply landscape. The tRNA gene names need to be properly
annotated for proper recognition. Expected format: tRNA-Asn-GTT-5-1.
}
\examples{
data(default_tTEscanR_tRNA_data)
tTEscanR_obj <- CreateObject(
    counts = default_tTEscanR_tRNA_data,
    assay = "tRNA"
)
tTEscanR_obj <- ComputeAnticodonUsage(object = tTEscanR_obj)
}
