% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonUsage.R
\name{computeAnticodonUsage}
\alias{computeAnticodonUsage}
\title{Compute Anticodon Usage from tRNA Gene Expression Data}
\usage{
computeAnticodonUsage(object, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
containing a tRNA assay.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites any existing assay and
metadata in the \code{object}. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
An updated \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
containing a new layer of information in the \code{assays}
representing the \code{"AnticodonUsage"}.
}
\description{
Calculates \strong{anticodon usage profiles} from tRNA gene
abundance data. It summarizes the abundance of tRNAs by their anticodon
identity, which can be used to estimate the tRNA supply landscape. The tRNA
gene names need to be properly annotated for proper recognition.
}
\examples{
data("default_tTEscanR_tRNA_data", package = "tTEscanR")

tTEscanR_obj <- createObject(
    counts = default_tTEscanR_tRNA_data,
    assay = "tRNA"
)
tTEscanR_obj <- computeAnticodonUsage(object = tTEscanR_obj)
}
