% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonUsage.R
\name{computeCodonUsage}
\alias{computeCodonUsage}
\title{Compute Codon Usage from mRNA Gene Expression Data}
\usage{
computeCodonUsage(
  object,
  codon_freq = NULL,
  species = NULL,
  additional_metrics = TRUE,
  reduce = 100,
  corr_method = "spearman",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
containing a mRNA assay.}

\item{codon_freq}{Optional; a user-provided codon frequency-per-gene table.
If necessary, it can be computed using \code{\link{getCodonFreq}}.}

\item{species}{Optional; either \code{"hg38"} (human) or \code{"mm39"}
(mouse) to load the default settings. Required if \code{codon_freq} is
not provided.}

\item{additional_metrics}{Logical; if \code{TRUE}, computes: (i) codon exonic
background, (ii) mean codon usage, and (iii) correlation between the
previous metrics. Defaults to \code{TRUE}.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression
values that exceed R's handling capacity. Defaults to 100.}

\item{corr_method}{A correlation method accepted by \code{\link{cor}}.
Required if \code{additional_metrics} is \code{TRUE}. Defaults to
\code{"spearman"}.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites any existing assay and
metadata in the \code{object}. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
An updated \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
containing a new layer of information in the \code{assays}
representing the \code{"CodonUsage"}. Additional computations will be
stored in the \code{param} slot as \code{"CodonUsage_AdditionalMetrics"}.
}
\description{
Estimates \strong{codon usage profiles} based on gene-level mRNA expression
data. It optionally accepts pre-computed codon frequency tables or uses
internally generated default tables when not provided. When enabled, it can
evaluate the correlation between background codon composition and observed
mean codon usage.
}
\examples{
data("default_tTEscanR_mRNA_data", package = "tTEscanR")
data("default_tTEscanR_metadata", package = "tTEscanR")

tTEscanR_obj <- createObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA",
    meta_data = default_tTEscanR_metadata,
    sample_id = "conditions",
    params = list("CorrectionFactor" = "tissue")
)
tTEscanR_obj <- computeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = FALSE, reduce = 1000
)
}
