% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonUsage.R
\name{ComputeCodonUsage}
\alias{ComputeCodonUsage}
\title{Compute Codon Usage from mRNA Gene Expression Data}
\usage{
ComputeCodonUsage(
  object,
  codon_freq = NULL,
  species = NULL,
  filter = "canonical",
  reduce = 100,
  additional.metrics = TRUE,
  corr_method = "spearman",
  overwrite.assay = FALSE,
  overwrite.metadata = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{tTEscanR_Object} containing a mRNA assay.}

\item{codon_freq}{Optional; a user-provided codon frequency-per-gene table. If necessary, it can be computed using \code{\link{ObtainCodonFreqPerGene}}.}

\item{species}{Optional; either \code{"hg38"} (human) or \code{"mm39"} (mouse) to load the default settings. Required if \code{codon_freq} is not provided.}

\item{filter}{Optional; either \code{"canonical"} (default) or \code{"length"} (longest transcript) to specify which transcript to choose if several are available for the same gene.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression values that exceed R's handling capacity. Defaults to 100.}

\item{additional.metrics}{Logical; if \code{TRUE}, computes: (i) codon exonic background, (ii) mean codon usage, and (iii) correlation between the previous metrics. Defaults to \code{TRUE}.}

\item{corr_method}{A correlation method accepted by \code{\link{cor}}. Required if \code{additional.metrics} is \code{TRUE}. Defaults to \code{"spearman"}.}

\item{overwrite.assay}{Logical; if \code{TRUE}, overwrites any existing assay in the \code{object}. Defaults to \code{FALSE}.}

\item{overwrite.metadata}{Logical; if \code{TRUE}, overwrites any existing metadata in the \code{object}. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
An updated \code{tTEscanR_Object} containing a new layer of information \code{"CodonUsage"} in the \code{assays} slot representing the codon usage.
Additional computations will be stored in the \code{meta.data} slot as \code{"CodonUsage_AdditionalMetrics"}.
}
\description{
This function estimates \strong{codon usage profiles} based on gene-level mRNA expression data stored in a \code{tTEscanR_Object}.
It optionally accepts pre-computed codon frequency tables or uses internally generated default tables when not provided.
When enabled, it can evaluate the correlation between background codon composition and observed mean codon usage.
If the additional metrics are to be computed the input \code{tTEscanR_Object} needs to have a \strong{"CorrectionFactor"} stored in the \code{"meta.data"} slot.
}
\examples{
data(subset_mRNA_data, metadata)
tTEscanR_obj <- Create_tTEscanR_Object(counts = subset_mRNA_data, assay = "mRNA",
                                       meta.data = list(metadata, "tissue"),
                                       meta.data.ids = list("ConditionsLabels", "CorrectionFactor"))
tTEscanR_obj <- ComputeCodonUsage(object = tTEscanR_obj, species = "hg38",
                                  additional.metrics = FALSE, reduce = 1000)
}
