% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_CodonUsage.R
\name{ComputeExonicBackground}
\alias{ComputeExonicBackground}
\title{Compute the Exonic Background of the Codon Usage}
\usage{
ComputeExonicBackground(data)
}
\arguments{
\item{data}{A codon usage matrix with codons as rows and conditions or samples as columns.}
}
\value{
A \code{matrix} with the codon exonic background.
}
\description{
This function calculates the \strong{codon usage background} based solely on exonic sequence composition, independent of gene expression levels.
It provides a reference distribution of codon frequencies across conditions, used to normalize/compare against observed codon usage patterns derived from expression data.
}
\examples{
data(subset_mRNA_data)
tTEscanR_obj <- Create_tTEscanR_Object(counts = subset_mRNA_data, assay = "mRNA")
tTEscanR_obj <- ComputeCodonUsage(object = tTEscanR_obj, species = "hg38",
                                  additional.metrics = FALSE, reduce = 1000)
exonic_background <- ComputeExonicBackground(data = tTEscanR_obj@assays$CodonUsage)
}
