% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_CodonUsage.R
\name{ComputeMeanUsage}
\alias{ComputeMeanUsage}
\title{Compute Usage Across Conditions (Mean Usage)}
\usage{
ComputeMeanUsage(
  data,
  assay = NULL,
  mode = "raw",
  metadata = NULL,
  corr_factor = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A \code{tTEscanR_Object} or expression count \code{matrix} (with codons, anticodons or amino acids as features).}

\item{assay}{Optional; a character string specifying the name of the assay to retrieve from the \code{tTEscanR_Object}.}

\item{mode}{Either \code{"raw"}, \code{"size-corrected"} or \code{"long-format"} to specify the format the input data belongs to. Defaults to \code{"raw"}.}

\item{metadata}{Optional; a \code{data.frame} with the meta-information related with the conditions in \code{data}. There has to be one column with the same labels as the column names.}

\item{corr_factor}{Optional; a factor based on \code{metadata} columns to define the variable to correct for. Required if \code{mode} is \code{"raw"} and \code{data} is not a \code{tTEscanR_Object}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
An updated \code{tTEscanR_Object} if \code{data} is a \code{tTEscanR_Object}. A \code{data.frame} containing a new layer of information representing the mean codon usage if \code{data} is an expression count matrix.
}
\description{
This function computes the \strong{average usage} of codon, anticodons, or amino acids across conditions, useful for summarizing feature usage trends across sample groups.
It supports direct input of a count matrix and extraction from a \code{tTEscanR_Object}.
When the input \code{data} is a \code{tTEscanR_Object} the parameters \code{metadata} and \code{corr_factor} will be extracted from the object, and ignored if specified as input parameters.
Therefore, variables \code{assay} and \code{metadata} need to be coherent with the rules described in \code{\link{Create_tTEscanR_Object}}.
}
\examples{
data(subset_tRNA_data, metadata)
tTEscanR_obj <- Create_tTEscanR_Object(counts = subset_tRNA_data, assay = "tRNA",
                                       meta.data = list(metadata, "tissue"),
                                       meta.data.ids = list("ConditionsLabels", "CorrectionFactor"))
# Input: tTEscanR object containing metadata and corr_factor parameters
tTEscanR_obj <- ComputeAnticodonUsage(object = tTEscanR_obj)
anticodon_mean_usage <- ComputeMeanUsage(data = tTEscanR_obj, assay = "AnticodonUsage")
}
