% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanRObject.R
\name{createObject}
\alias{createObject}
\title{Create a tTEscanR object as a MultiAssayExperiment}
\usage{
createObject(
  counts,
  assay = "mRNA",
  meta_data = NULL,
  sample_id = NULL,
  params = list(),
  verbose = TRUE
)
}
\arguments{
\item{counts}{A matrix, data.frame or list of count matrices.}

\item{assay}{Optional; a character string or vector specifying the assay
name(s) when \code{counts} is a matrix, data.frame or unnamed list.
Defaults to \code{"mRNA"}.}

\item{meta_data}{Optional; a data.frame of sample metadata.}

\item{sample_id}{Optional; a character string naming the column in
\code{meta.data} to use as sample row names.}

\item{params}{Optional; a list of global run parameters.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
A \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} object.
}
\description{
Initializes a \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} object
to store translation efficiency data layers.
}
\examples{
data("default_tTEscanR_mRNA_data", package = "tTEscanR")
data("default_tTEscanR_tRNA_data", package = "tTEscanR")
data("default_tTEscanR_metadata", package = "tTEscanR")

tTEscanR_obj <- createObject(
    counts = list(
        mRNA = default_tTEscanR_mRNA_data,
        tRNA = default_tTEscanR_tRNA_data
    ),
    meta_data = default_tTEscanR_metadata,
    sample_id = "conditions"
)

}
