% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_tTEscanR_mRNA_data}
\alias{default_tTEscanR_mRNA_data}
\title{mRNA Expression Data Subset}
\format{
A matrix where rows represent genes and columns represent individual
samples or cell-types.
}
\source{
Subset of preprocessed count table generated by Gao et al. (2022),
derived from raw sequencing data originally published by Cao et
al. (2020).
}
\usage{
data(default_tTEscanR_mRNA_data)
}
\description{
A subset of preprocessed mRNA gene expression count data used in package
vignettes and examples.
}
\references{
Gao W, Gallardo-Dodd CJ, Kutter C (2022). "Cell type-specific analysis by
single-cell profiling identifies a stable mammalian tRNA-mRNA interface and
increased translation efficiency in neurons." \emph{Genome Research},
32(1), 97–110. \doi{10.1101/gr.275944.121}

Cao J, O'Day DR, Pliner HA, et al. (2020). "A human cell atlas of fetal
gene expression." \emph{Science}, 370(6518), eaba7721.
\doi{10.1126/science.aba7721}
}
\keyword{datasets}
