% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_tTEscanR_tRNA_data}
\alias{default_tTEscanR_tRNA_data}
\title{tRNA Expression Data Subset}
\format{
A matrix or data frame where rows represent tRNA gene names
and columns represent individual samples or cell types.
}
\usage{
data(default_tTEscanR_tRNA_data)
}
\description{
This dataset contains an example of tRNA gene expression data with tRNA
genes as rows and cell types as columns.
}
\keyword{datasets}
