% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionAnalysis.R
\name{ExecuteDESeq2runner}
\alias{ExecuteDESeq2runner}
\title{Perform Differential Expression Analysis Using DESeq2}
\usage{
ExecuteDESeq2runner(
  list_data,
  metadata,
  targets = NULL,
  target_factor = NULL,
  corr_factor = NULL,
  fc_threshold = 1,
  pval_threshold = 0.05,
  reduce = 100,
  heatmap = TRUE,
  PCA = TRUE,
  numPC = 2,
  color_factor = NULL,
  shape_factor = NULL,
  color_palette = NULL,
  labels = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{list_data}{A \code{list} of matrices with features (i.e. genes, codons, anticodons or amino acids) as rows and samples or conditions as columns. The list can be named or unnamed.}

\item{metadata}{A \code{data.frame} with the data associated to the conditions in the matrices of \code{list_data}. There has to be one column with the same labels as the column names.}

\item{targets}{Optional; a \code{data.frame} with one column indicating the conditions to select, and another column with the labels for the comparisons.}

\item{target_factor}{Optional; a factor based on \code{metadata} columns to define the variable to classify the entries in \code{targets}. Required if \code{targets} provided.}

\item{corr_factor}{Optional; name of the categorical variable to correct the data. Required if \code{PCA} is \code{TRUE}.}

\item{fc_threshold}{Numeric; fold change threshold used for highlighting significant features in the volcano plot (if \code{targets} is specified). Defaults to 1.}

\item{pval_threshold}{Numeric; p-value threshold used for highlighting significant features in the volcano plot. Required if \code{targets} is specified. Defaults to 0.05.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression values that exceed R's handling capacity. Defaults to 100.}

\item{heatmap}{Logical; if \code{TRUE}, generates a heatmap for exploratory analysis. Not applicable if \code{targets} is specified. Defaults to \code{TRUE}.}

\item{PCA}{Logical; if \code{TRUE}, generates a principal component analysis. Not applicable if \code{targets} is specified. Defaults to \code{TRUE}.}

\item{numPC}{Numeric; number of principal components to include in the PCA analysis. Required if \code{PCA} is \code{TRUE}. Defaults to 2.}

\item{color_factor}{Optional; name of the categorical variable to group by color the data points. Required if \code{PCA} is \code{TRUE}.}

\item{shape_factor}{Optional; name of the categorical variable to group by shape the data points. Required if \code{PCA} is \code{TRUE}.}

\item{color_palette}{Optional; a \code{vector} of color codes to customize PCA plot appearance.}

\item{labels}{Logical; if \code{TRUE} includes the data points labels in the PCA plot. Required if \code{PCA} is \code{TRUE}. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
A \code{list} of outputs per each matrix in \code{list_data}, based on the enabled parameters: (i) heatmap, (ii) PCA plot, and (iii) Size corrected \code{data}.
}
\description{
This function applies differential expression analysis using the DESeq2 framework on a matrix of expression values.
It supports both exploratory visualizations (heatmap and PCA) and targeted comparisons using a custom contrast table.
}
\examples{
data(subset_mRNA_data, metadata)
DE_analysis <- ExecuteDESeq2runner(list_data = list(mRNA = subset_mRNA_data),
                                   metadata = metadata, corr_factor = "tissue")
}
