% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTE_HelperFunctions_GeneralInternal.R
\name{featuresToAA}
\alias{featuresToAA}
\title{Relates Codons, Anticodons and their corresponding Amino Acids}
\usage{
featuresToAA(
  data,
  position = NULL,
  genetic_code = "Standard",
  verbose = TRUE,
  notation_from = c("codon", "anticodon"),
  notation_to = c("aa", "anticodon", "codon")
)
}
\arguments{
\item{data}{A \code{character} vector or \code{data.frame}
containing gene names or features to be translated.}

\item{position}{Optional; either \code{"row"}, \code{"column"} or
\code{<column_name>} to specify the location of the genes or features in
\code{data}. Required if \code{data_to_transalte} is a
\code{data.frame}.}

\item{genetic_code}{A \code{character} string to specify the genetic code to
be used. Defaults to \code{"Standard"}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}

\item{notation_from}{Either \code{"codon"} or \code{"anticodon"} to select
the input format of the features in \code{data_to_translate}.}

\item{notation_to}{Either \code{"codon"}, \code{"anticodon"} or \code{"aa"}
to select the output format of the features in \code{data}.}
}
\value{
Translated features (codons, anticodons or amino acids) from
\code{data_to_translate}.
}
\description{
This function converts: codons and anticodons into anticodons, codons or
amino acids based on the genetic code.
}
\examples{
data(default_tTEscanR_mRNA_data, default_tTEscanR_metadata)
tTEscanR_obj <- createObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA",
    meta_data = default_tTEscanR_metadata,
    sample_id = "conditions",
    params = list("CorrectionFactor" = "tissue")
)
tTEscanR_obj <- computeCodonUsage(
    object = tTEscanR_obj, codon_freq = NULL,
    species = "hg38", additional_metrics = FALSE
)
codons <- rownames(SummarizedExperiment::assay(tTEscanR_obj, "CodonUsage"))
codons_to_AA <- featuresToAA(
    data = codons, notation_from = "codon", notation_to = "aa"
)
codons_to_anticodons <- featuresToAA(
    data = codons,
    notation_from = "codon", notation_to = "anticodon"
)
}
