% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanRObject.R
\name{getAssay}
\alias{getAssay}
\alias{getAssay,tTEscanR_Object-method}
\title{Get Assay Data from a tTEscanR Object}
\usage{
getAssay(object, name)

\S4method{getAssay}{tTEscanR_Object}(object, name)
}
\arguments{
\item{object}{A \code{tTEscanR_Object}.}

\item{name}{A character string specifying the name of the assay to retrieve
(e.g. "mRNA", "tRNA").}
}
\value{
The requested assay data (typically a matrix or data.frame).
}
\description{
This function safely retrieves the specified data from a
\code{tTEscanR_Object}.
}
\examples{
data(default_tTEscanR_mRNA_data)
tTEscanR_obj <- CreateObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA"
)
mRNA_data <- getAssay(tTEscanR_obj, "mRNA")
}
