% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralProcessing.R
\name{MergeMatrices}
\alias{MergeMatrices}
\title{Combine large matrices}
\usage{
MergeMatrices(...)
}
\arguments{
\item{...}{A variable number of \code{matrix}.}
}
\value{
A single sparse matrix with as a combination of all the input
matrices.
}
\description{
This function efficiently combines individual matrices.
}
\examples{
df1 <- matrix(c(1, 0, 0, 2), nrow = 2, dimnames = list(
    c("geneA", "geneB"),
    c("s1", "s2")
))
df2 <- matrix(c(3, 0, 0, 4), nrow = 2, dimnames = list(
    c("geneB", "geneC"),
    c("s2", "s3")
))
merged_matrix <- MergeMatrices(df1, df2)
}
