% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Correlation Plot: Exonic Codon Background - Mean Codon Usage}
\usage{
plotCorrelation(
  data,
  plot = "MeanCodonUsage",
  x_axis_col,
  y_axis_col,
  condition_col,
  extra_val = NULL,
  label_col = NULL,
  color_palette = NULL,
  filename = NULL,
  show_legend = "none",
  targeted_arg = NULL,
  add_titles = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A long format table. This format can be obtained using
\code{\link{transformFormat}}.}

\item{plot}{Either \code{"MeanCodonUsage"} (default) or
\code{"PoolDiversity"} to indicate the \code{data} source.}

\item{x_axis_col}{Name of the categorical variable to reflect in the plot.}

\item{y_axis_col}{Name of the numerical variable to reflect in the plot.}

\item{condition_col}{Name of the categorical variable to group the data
points.}

\item{extra_val}{Optional; variable with additional information to include
in the plot (e.g. correlation value).}

\item{label_col}{Name of the categorical variable to label the data points.}

\item{color_palette}{Optional; a vector of color codes to customize plot
appearance.}

\item{filename}{Optional; name for the plot to be saved.}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the
position of the legend.}

\item{targeted_arg}{Optional; a vector defining key feature clusters to
highlight or label.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot.
Defaults to \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the correlation. If
\code{save_format} is provided, the plot will also be saved to the
specified location.
}
\description{
This function generates a visualization to correlate different parameters of
the \code{data}. The input data (\code{data}) is expected to be in long
format and contain a minimum information regarding the features, the
conditions they belong to and their usage counts. A common usage of this
kind of plot is to represent the codon frequencies correlations of the
exonic background and the mean codon usage across conditions. Based on the
selected \code{plot} parameter the user can select the most convenient layout
to display the data. It is crucial to ensure consistency between the name of
the columns in \code{data} and the parameters \code{x_axis_col},
\code{y_axis_col} and \code{condition_col}.
}
\examples{
data("default_tTEscanR_mRNA_data", package = "tTEscanR")
data("default_tTEscanR_metadata", package = "tTEscanR")

# Define the object and compute the codon usage
tTEscanR_obj <- createObject(
    counts = default_tTEscanR_mRNA_data, assay = "mRNA",
    meta_data = default_tTEscanR_metadata,
    sample_id = "conditions",
    params = list("CorrectionFactor" = "tissue")
)
tTEscanR_obj <- computeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = TRUE, reduce = 1000
)

# Compute and extract the mean codon usage
meta <- S4Vectors::metadata(tTEscanR_obj)
additional_metrics <- meta[["CodonUsage_AdditionalMetrics"]]
mean_codon_usage <- additional_metrics$MeanCodonUsage
exonic_background <- additional_metrics$CodonExonicBackground
exonic_background <- as.data.frame(exonic_background)
correlation_mean_background <- cbind(mean_codon_usage, exonic_background)

plotCorrelation(
    data = correlation_mean_background, plot = "MeanCodonUsage",
    x_axis_col = "mean_usage_across_conditions",
    y_axis_col = "exonic_background", condition_col = "feature",
    extra_val = additional_metrics$MeanCodonCorr, add_titles = TRUE,
    show_legend = "none"
)
}
