% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionAnalysis.R
\name{PlotDEResults}
\alias{PlotDEResults}
\title{Generates Visualizations from the DEA data}
\usage{
PlotDEResults(
    DE_results_list,
    dataset_name = NULL,
    heatmap = TRUE,
    dim_reduct = NULL,
    numPC = 2,
    condition = NULL,
    verbose = TRUE,
    color_factor = condition,
    shape_factor = NULL,
    label_factor = NULL,
    highlight_median = FALSE,
    scale_pca = FALSE,
    color_palette = NULL,
    fc_threshold = 1,
    padj_threshold = 0.05,
    label_significant = TRUE,
    show_legend = "none"
)
}
\arguments{
\item{DE_results_list}{A DESeq2 object with the normalized and vst counts.
Can be obtained by running \code{\link{ComputeDEResults}}}

\item{dataset_name}{String to specify the assay in \code{DE_results_list}
to extract.}

\item{heatmap}{Logical; if \code{TRUE}, generates a heatmap for exploratory
analysis. Defaults to \code{TRUE}.}

\item{dim_reduct}{Either \code{"PCA"}, \code{"UMAP"} or \code{"tSNE"} to
specify the dimensionality reduction approach to be executed. Defaults to
\code{"PCA"}.}

\item{numPC}{Numeric; number of principal components to include in the PCA
analysis. Required if \code{dim_reduct} is \code{"PCA"}. Defaults to 2.}

\item{condition}{Optional; a factor based on \code{metadata} columns to
define the comparisons to perform in a targeted analysis. Defaults to
\code{NULL}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}

\item{color_factor}{Optional; name of the categorical variable in
\code{metadata} to group by color the data points. Used if \code{dim_reduct}
specified.}

\item{shape_factor}{Optional; name of the categorical variable to label the
data points. Used if \code{dim_reduct} specified.}

\item{label_factor}{Optional; name of the categorical variable to label the
data points. Used if \code{dim_reduct} specified.}

\item{highlight_median}{Logical; if \code{TRUE} the data points of each
cluster will be summarized into the median. Defaults to \code{FALSE}.}

\item{scale_pca}{Logical; if \code{TRUE}, scales the data if
\code{dim_reduct} is \code{"PCA"}. Defaults to \code{FALSE}.}

\item{color_palette}{Optional; a \code{vector} of color codes to customize
the plot appearance.}

\item{fc_threshold}{Numeric; fold change threshold used for highlighting
significant features in the volcano plot (if \code{targets} is specified).
Defaults to 1.}

\item{padj_threshold}{Numeric; p-value threshold used for highlighting
significant features in the volcano plot. Defaults to 0.05.}

\item{label_significant}{Logical; if \code{TRUE} displays the axis of the
plots based on \code{fc_threshold} and \code{padj_threshold}. Defaults to
\code{TRUE}.}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of
the legend in the plot.}
}
\value{
Visualization of the DEA results.
}
\description{
Generates Visualizations from the DEA data
}
\examples{
data(default_tTEscanR_tRNA_data, default_tTEscanR_metadata)
DE_analysis <- ComputeDEResults(
    list_data = list(tRNA = default_tTEscanR_tRNA_data),
    metadata = default_tTEscanR_metadata, batch = "tissue"
)
DE_plots <- PlotDEResults(
    DE_results_list = DE_analysis, dataset_name = "tRNA",
    dim_reduct = "PCA", condition = "tissue", heatmap = FALSE
)
}
