% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Distribution Plot of Codon/Anticodon Usage or Amino Acid Demand/Supply}
\usage{
plotDistribution(
  data,
  plot = "jitter",
  bar_position = "dodge",
  x_axis_col,
  y_axis_col,
  condition_col,
  color_palette = NULL,
  ncols = 1,
  facet_col = NULL,
  add_stats = FALSE,
  targeted_arg = NULL,
  save_format = NULL,
  out_name = NULL,
  out_directory = NULL,
  show_legend = "none",
  add_titles = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A long format table. This format can be obtained using
\code{\link{transformFormat}}.}

\item{plot}{Either \code{"jitter"} (default), \code{"barplot"},
\code{"boxplot"} or \code{"dot"} to indicate the type of plot to
generate.}

\item{bar_position}{Either \code{"dodge"} (default), \code{"stack"} or
\code{"fill"} to indicate the display of the plot if \code{plot} is
\code{"barplot"}.}

\item{x_axis_col}{Name of the categorical variable to reflect in the plot.}

\item{y_axis_col}{Name of the numerical variable to reflect in the plot.}

\item{condition_col}{Name of the categorical variable to group the data
points.}

\item{color_palette}{Optional; a vector of color codes to customize plot
appearance.}

\item{ncols}{Numeric; number of columns for arranging panels. Defaults to 1.}

\item{facet_col}{Optional; name of the categorical variable to divide the
plot into different panels. Required if \code{ncols} bigger than 1.}

\item{add_stats}{Logical; if \code{TRUE}, performs a statistical analysis
based on the available parameters. Defaults to \code{FALSE}.}

\item{targeted_arg}{Optional; a vector defining key feature clusters to
highlight or label.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify
the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format}
specified).}

\item{out_directory}{Optional; path to the directory where the plot will
be saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the
position of the legend.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot.
Defaults to \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the requested distribution plot.
If \code{save_format} is provided, the plot will also be saved to the
specified location.
}
\description{
This function generates a visualization of codon-anticodon usage or amino
acid demand-supply distributions across conditions. The input data
(\code{data}) is expected to be in long format and contain a minimum
information regarding the features, the conditions they belong to and their
usage counts. Based on the selected \code{plot} parameter the user can
select the most convenient layout to display the data. It is crucial to
ensure consistency between the name of the columns in \code{data} and the
parameters \code{x_axis_col}, \code{y_axis_col} and \code{condition_col}.
}
\examples{
data(default_tTEscanR_mRNA_data, default_tTEscanR_metadata)

# Define the object and compute the codon usage
tTEscanR_obj <- createObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA",
    meta.data = list(default_tTEscanR_metadata, "tissue"),
    meta.data.ids = list("ConditionsLabels", "CorrectionFactor")
)
tTEscanR_obj <- computeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = FALSE, reduce = 1000
)

# Transform the data
long_cu <- transformFormat(
    data = getAssay(tTEscanR_obj, "CodonUsage"), normalize = TRUE,
    rownames_to_column = "codon", names_to = "condition", values_to = "usage"
)
long_cu <- long_cu |>
    tidyr::separate(condition, into = c("tissue", "cell_type"), sep = "-")

# Generate the plot
codon_usage_plot <- plotDistribution(
    data = long_cu, plot = "jitter", x_axis_col = "codon",
    y_axis_col = "usage", condition_col = "tissue", show_legend = "right",
    add_titles = FALSE
)
}
