% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{PlotProportion}
\alias{PlotProportion}
\title{Proportion Plot of Codon/Anticodon Usage or Amino Acid Demand/Supply}
\usage{
PlotProportion(
    data,
    plot = "bar",
    var_numerical,
    var_categorical,
    var_color = NULL,
    facet_col = NULL,
    color_palette = NULL,
    num_limits = NULL,
    num_rings = 5,
    save_format = NULL,
    out_name = NULL,
    out_directory = NULL,
    zoom = FALSE,
    show_legend = "none",
    add_titles = TRUE,
    order = NULL,
    normalize = TRUE,
    verbose = TRUE
)
}
\arguments{
\item{data}{A long format table. This format can be obtained using
\code{\link{TransformFormat}}.}

\item{plot}{Either \code{"bar"} (default), \code{"donut"} or \code{"radar"}
to indicate the type of plot to generate.}

\item{var_numerical}{Name of the numerical variable to reflect in the plot.}

\item{var_categorical}{Name of the categorical variable to reflect in the
plot.}

\item{var_color}{Optional; name of the categorical variable to group the
data points when coloring. Required if \code{plot} is \code{"bar"}.}

\item{facet_col}{Optional; name of the categorical variable to divide the
plot into different panels.}

\item{color_palette}{Optional; a vector of color codes to customize plot
appearance.}

\item{num_limits}{Optional; a vector with the upper and lower ranges of the
values in \code{var_numerical}.}

\item{num_rings}{Optional; a number specifying the amount of rings to
display if \code{plot} is \code{"radar"}. Defaults to 5.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify
the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format}
specified).}

\item{out_directory}{Optional; path to the directory where the plot will be
saved (if \code{save_format} specified).}

\item{zoom}{Logical; if \code{TRUE}, centers the plot display to the values
in \code{data}.Defaults to \code{TRUE}.}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of
the legend.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot.
Defaults to \code{TRUE}.}

\item{order}{Optional; a vector of the levels to organize the data, based
on the \code{var_categorical}.}

\item{normalize}{Logical; if \code{TRUE}, normalizes the data in order to
display the relative contribution of the \code{var_categorical}.
Defaults to \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the requested proportion plot.
If \code{save_format} is provided, the plot will also be saved to the
specified location.
}
\description{
This function generates a proportion plot to compare codon-anticodon usage
or amino acid demand-supply frequencies across conditions. The input data
(\code{data}) is expected to be in long format and contain a minimum
information regarding the features, the conditions they belong to and their
usage counts. Based on the selected \code{plot} parameter the user can
select the most convenient layout to display the data. It is crucial to
ensure consistency between the name of the columns in \code{data} and the
parameters \code{var_numerical}, \code{var_categorical} and \code{var_color}.
}
\examples{
data(default_tTEscanR_mRNA_data, default_tTEscanR_metadata)

# Define the object and compute the codon usage
tTEscanR_obj <- CreateObject(
    counts = default_tTEscanR_mRNA_data, assay = "mRNA",
    meta.data = list(default_tTEscanR_metadata, "tissue"),
    meta.data.ids = list("ConditionsLabels", "CorrectionFactor")
)
tTEscanR_obj <- ComputeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = TRUE, reduce = 1000
)

# Compute and extract the mean codon usage
additional_metrics <- getMetadata(
    tTEscanR_obj, "CodonUsage_AdditionalMetrics"
)
mean_codon_usage <- additional_metrics$MeanCodonUsage
mean_codon_usage$codon <- mean_codon_usage$feature

# Translate the codons to amino acids
mean_codon_usage <- FeaturesToAA(
    data = mean_codon_usage, position = "feature",
    notation_from = "codon", notation_to = "aa", verbose = FALSE
)

# Generate the plot
PlotProportion(
    data = mean_codon_usage, plot = "bar",
    var_numerical = "mean_usage_across_conditions",
    var_categorical = "codon", var_color = "feature", show_legend = "none"
)
}
