% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{PlotTEscore}
\alias{PlotTEscore}
\title{Violin Plot Displaying tTE Scores}
\usage{
PlotTEscore(
    data,
    metadata,
    class_col,
    index_col,
    target_col = NULL,
    facet_col = NULL,
    color_palette = NULL,
    save_format = NULL,
    out_name = NULL,
    add_stats = TRUE,
    out_directory = NULL,
    show_legend = "none",
    add_titles = TRUE,
    show_outliers = FALSE,
    verbose = TRUE
)
}
\arguments{
\item{data}{A tTE results table obtained from \code{\link{Compute_tTE}}.}

\item{metadata}{A table with additional information regarding the conditions
in \code{data}.}

\item{class_col}{Name of the categorical variable to reflect in the plot.}

\item{index_col}{Name of the categorical variable that links the conditions
in \code{data} with the \code{metadata}.}

\item{target_col}{Optional; name of the categorical variable to perform the
statistical comparison (the most specific level). Used if \code{add_stats}
is \code{TRUE}.}

\item{facet_col}{Optional; name of the categorical variable separate the
plot into different panels.}

\item{color_palette}{Optional; a vector of color codes to customize plot
appearance.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify
the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format}
specified).}

\item{add_stats}{Logical; if \code{TRUE}, performs a statistical analysis
based on the available parameters. Defaults to \code{TRUE}.}

\item{out_directory}{Optional; path to the directory where the plot will be
saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"},
\code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of
the legend in the plot.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot.
Defaults to \code{TRUE}.}

\item{show_outliers}{Logical; if \code{TRUE}, labels the outlier data points
based on \code{index_col}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the tTE scores. If
\code{save_format} is provided, the plot will also be saved to the specified
location. If \code{add_stats} reports a table with the statitical measures
summarized.
}
\description{
This function generates a violin plot to visualize the distribution of tTE
scores across different condition. The tTE scores should be obtained using
\code{\link{Compute_tTE}}. The plot allows the user to easily compare the
tTE distribution between different conditions, with the option to highlight
specific feature clusters based on the provided \code{targets}.
}
\examples{
data(
    default_tTEscanR_mRNA_data, default_tTEscanR_tRNA_data,
    default_tTEscanR_metadata
)

# Define the tTEscanR object
tTEscanR_obj <- CreateObject(
    counts = list(
        mRNA = default_tTEscanR_mRNA_data, tRNA = default_tTEscanR_tRNA_data
    ),
    meta.data = list(default_tTEscanR_metadata, "tissue"),
    meta.data.ids = list("ConditionsLabels", "CorrectionFactor")
)

# Compute the codon and anticodon usage
tTEscanR_obj <- ComputeCodonUsage(
    object = tTEscanR_obj, species = "hg38",
    additional_metrics = FALSE, reduce = 10000
)
tTEscanR_obj <- ComputeAnticodonUsage(object = tTEscanR_obj)

# Compute the theoretical translation efficiency (tTE scores)
tTEscanR_obj <- Compute_tTE(
    object = tTEscanR_obj, level = "codon", compute_significance = TRUE
)
tTEresults_codon <- getMetadata(tTEscanR_obj, "tTEresults_codon")
conditions_metadata <- getMetadata(tTEscanR_obj, "ConditionsLabels")

# Visualize the tTE scores
PlotTEscore(
    data = tTEresults_codon, metadata = conditions_metadata,
    index_col = "conditions", class_col = "tissue", add_stats = TRUE
)
}
