% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanR_General.R
\name{Run_tTEscanR_pipeline}
\alias{Run_tTEscanR_pipeline}
\title{Runs the Theoretical Translation Efficiency (tTE) Pipeline}
\usage{
Run_tTEscanR_pipeline(
  mRNA_data,
  tRNA_data,
  metadata,
  corr_factor = NULL,
  corr_method = "spearman",
  additional.metrics = TRUE,
  compute.significance = TRUE,
  codon_freq = NULL,
  species = NULL,
  filter = "canonical",
  reduce = 100,
  runDESeq = TRUE,
  color_factor = NULL,
  verbose = TRUE
)
}
\arguments{
\item{mRNA_data}{A count matrix of mRNA genes (rows) per conditions (columns).}

\item{tRNA_data}{A count matrix of tRNA genes (rows) per conditions (columns).}

\item{metadata}{A \code{data.frame} with the meta-information related with the conditions in \code{mRNA_data} or \code{tRNA_data}. Just the intersecting conditions will be considered.}

\item{corr_factor}{A factor based on \code{metadata} columns to define the variable to correct for when size correcting the count matrices.}

\item{corr_method}{A correlation method accepted by \code{\link{cor}}. Defaults to \code{"spearman"}.}

\item{additional.metrics}{Logical; if \code{TRUE}, computes: (i) codon exonic background, (ii) mean codon usage, and (iii) correlation between the previous metrics. Defaults to \code{TRUE}.}

\item{compute.significance}{Logical; if \code{TRUE}, computes the statistical significance (p-value) of the tTE scores. Defaults to \code{TRUE}.}

\item{codon_freq}{Optional; a user-provided codon frequency-per-gene table. If necessary, it can be computed using \code{\link{ObtainCodonFreqPerGene}}.}

\item{species}{Either \code{"hg38"} (human) or \code{"mm39"} (mouse) to specify which default codon frequency-per-gene table to use. Required if \code{codon_freq} not provided or if the gene annotation is inconsistent between both inputs.}

\item{filter}{Either \code{"canonical"} (default) or \code{"length"} (longest transcript) to specify which transcript to choose if several are available for the same gene.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression values that exceed R's handling capacity. Defaults to 100.}

\item{runDESeq}{Logical; if \code{TRUE}, performs differential expression analysis to each assay in the \code{tTEscanR_Object}. Defaults to \code{TRUE}}

\item{color_factor}{A factor based on \code{metadata} columns to define the colors in the PCA plots. Required if \code{runDESeq} is \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}
}
\value{
A \code{tTEscanR_Object} with the assays and metadata computed through the tTE pipeline.
}
\description{
This function wraps up all the independent functions of theoretical translation efficiency pipeline.
Requires an mRNA and tRNA count matrices to compute the codon and anticodon usage and further derive the amino acid demand and supply.
With matching condition in the former matrices the theoretical translation efficiency would be computed.
}
\examples{
data(subset_mRNA_data, subset_tRNA_data, metadata)
tTEscanR_obj <- Run_tTEscanR_pipeline(mRNA_data = subset_mRNA_data, tRNA_data = subset_tRNA_data,
                                      metadata = metadata, species = "hg38",
                                      corr_factor = "tissue", additional.metrics = FALSE,
                                      compute.significance = FALSE, runDESeq = FALSE)
}
