% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionAnalysis.R
\name{RunDEAnalysis}
\alias{RunDEAnalysis}
\title{Perform Differential Expression Analysis Using DESeq2}
\usage{
RunDEAnalysis(
  list_data,
  metadata,
  condition = NULL,
  batch = NULL,
  reference = NULL,
  reduce = 100,
  dim_reduct = NULL,
  color_factor = NULL,
  heatmap = TRUE,
  shape_factor = NULL,
  label_factor = NULL,
  highlight_median = FALSE,
  numPC = 2,
  color_palette = NULL,
  fc_threshold = 1,
  padj_threshold = 0.05,
  label_significant = TRUE,
  verbose = TRUE,
  show_legend = "none"
)
}
\arguments{
\item{list_data}{A \code{list} of matrices with features (i.e. genes, codons, anticodons or amino acids) as rows and samples or conditions as columns. The list can be named or unnamed.}

\item{metadata}{A \code{data.frame} with the data associated to the conditions in the matrices of \code{list_data}. There has to be one column with the same labels as the column names.}

\item{condition}{Optional; a factor based on \code{metadata} columns to define the comparisons to perform in a targeted analysis. Defaults to \code{NULL}.}

\item{batch}{Optional; name of the categorical variable in \code{metadata} to correct the data.}

\item{reference}{Optional; factor from the \code{batch} variable to use as reference for the corrections. If not specified, the 1st factor that appears will be used instead.}

\item{reduce}{Numeric; a scaling factor used to normalize large expression values that exceed R's handling capacity. Defaults to 100.}

\item{dim_reduct}{Either \code{"PCA"}, \code{"UMAP"} or \code{"tSNE"} to specify the dimensionality reduction approach to be executed. Defaults to \code{"PCA"}.}

\item{color_factor}{Optional; name of the categorical variable in \code{metadata} to group by color the data points. Used if \code{dim_reduct} specified.}

\item{heatmap}{Logical; if \code{TRUE}, generates a heatmap for exploratory analysis. Not applicable if \code{targets} is specified. Defaults to \code{TRUE}.}

\item{shape_factor}{Optional; name of the categorical variable in \code{metadata} to group by shape the data points. Used if \code{dim_reduct} specified.}

\item{label_factor}{Optional; name of the categorical variable to label the data points. Used if \code{dim_reduct} specified.}

\item{highlight_median}{Logical; if \code{TRUE} the data points of each cluster will be summarized into the median. Defaults to \code{FALSE}.}

\item{numPC}{Numeric; number of principal components to include in the PCA analysis. Required if \code{dim_reduct} is \code{"PCA"}. Defaults to 2.}

\item{color_palette}{Optional; a \code{vector} of color codes to customize the plot appearance.}

\item{fc_threshold}{Numeric; fold change threshold used for highlighting significant features in the volcano plot (if \code{targets} is specified). Defaults to 1.}

\item{padj_threshold}{Numeric; p-value threshold used for highlighting significant features in the volcano plot. Defaults to 0.05.}

\item{label_significant}{Logical; if \code{TRUE} displays the axis of the plots based on \code{fc_threshold} and \code{padj_threshold}. Defaults to \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages. Defaults to \code{TRUE}.}

\item{show_legend}{Either \code{"none"} (default), \code{"top"}, \code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of the legend in the plot.}
}
\value{
A \code{list} of outputs per each matrix in \code{list_data}, based on the enabled parameters: (i) exploratory plots (heatmap and/or PCA), (ii) targeted plot (volcano), and (iii) size-corrected \code{data}.
}
\description{
This function applies differential expression analysis using the DESeq2 framework on a matrix of expression values.
It supports both exploratory visualizations (heatmap and PCA) and targeted comparisons using a custom contrast table.
}
\examples{
data(default_tTEscanR_tRNA_data, default_tTEscanR_metadata)
DE_analysis <- RunDEAnalysis(list_data = list(mRNA = default_tTEscanR_tRNA_data),
                             metadata = default_tTEscanR_metadata, batch = "tissue",
                             color_factor = "tissue")
}
