% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralProcessing.R
\name{Set_tRNAgenes}
\alias{Set_tRNAgenes}
\title{Translate chromatin coordinates to tRNA genes}
\usage{
Set_tRNAgenes(data, tRNA_bed, flanking_region = 100)
}
\arguments{
\item{data}{A tRNA expression matrix with tRNA genes as rows and samples or conditions as columns. Can be generated using \code{\link{Get_tRNAMatrix}}}

\item{tRNA_bed}{A string indicating the path to the \code{".bed"} file from the organism's reference genome}

\item{flanking_region}{Integer; number of nucleotides that form the flanking region of each tRNA. Defaults to 100.}
}
\value{
The \code{data} tRNA expression data with the tRNA genes annotated for the given fragment coordinates.
}
\description{
Translate chromatin coordinates to tRNA genes
}
