% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subset_tRNA_data}
\alias{subset_tRNA_data}
\title{tRNA Expression Data Subset
This dataset contains a subset of tRNA_data.}
\usage{
data(subset_tRNA_data)
}
\description{
tRNA Expression Data Subset
This dataset contains a subset of tRNA_data.
}
