% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralProcessing.R
\name{tRNAGetMatrix}
\alias{tRNAGetMatrix}
\title{Generate a tRNA expression matrix}
\usage{
tRNAGetMatrix(
    data,
    assay = "peaks",
    confidence_set = NULL,
    tRNA_name_map = NULL,
    species = NULL,
    flanking_region = 100,
    name_sep = c("-", "-"),
    save = TRUE,
    out_name = NULL,
    out_directory = NULL,
    verbose = TRUE
)
}
\arguments{
\item{data}{\code{SummarizedExperiment}, \code{ChromatinAssay}, or
\code{SeuratObject}.}

\item{assay}{Optional; a character string specifying the name of the assay to
retrieve from \code{data} if it is a \code{SeuratObject}. Defaults to
\code{"peaks"}.}

\item{confidence_set}{Either a file path to the tRNA annotations
(confidence set file from gtRNAdb), or a GRanges object. Contains the set of
high confidence tRNA genes}

\item{tRNA_name_map}{Optional; a \code{data.frame} with the tRNA gene names
linked to \code{confidence_set}. Contains two columns: tRNAscan-SE and
GtRNAdb gene ids.}

\item{species}{Optional; either \code{"hg38"} (human) or \code{"mm39"}
(mouse) to load the default \code{confidence_set}}

\item{flanking_region}{Integer; number of nucleotides that form the flanking
region of each tRNA. Defaults to 100.}

\item{name_sep}{A string delimiter to format the tRNA gene names in the
output matrix. Defaults to \code{c("-", "-")}.}

\item{save}{Logical; if \code{TRUE} stores the generated tRNA matrix into a
file.}

\item{out_name}{Optional; name for the saved plot (if \code{save} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be
saved (if \code{save} specified).}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
Sparse matrix of tRNA counts (tRNAs x cells)
}
\description{
Generate a tRNA expression matrix
}
