% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{tTE_DistributionPlot}
\alias{tTE_DistributionPlot}
\title{Distribution Plot of Codon/Anticodon Usage or Amino Acid Demand/Supply}
\usage{
tTE_DistributionPlot(
  data,
  plot = "jitter",
  bar_position = "dodge",
  x_axis_col,
  y_axis_col,
  condition_col,
  ncols = 1,
  facet_col = NULL,
  add_stats = FALSE,
  color_palette = NULL,
  targeted_arg = NULL,
  save_format = NULL,
  out_name = NULL,
  out_directory = NULL,
  show_legend = "none",
  add_titles = TRUE
)
}
\arguments{
\item{data}{A long format table. This format can be obtained using \code{\link{TransformFormat}}.}

\item{plot}{Either \code{"jitter"} (default), \code{"barplot"}, \code{"boxplot"} or \code{"dot"} to indicate the type of plot to generate.}

\item{bar_position}{Either \code{"dodge"} (default), \code{"stack"} or \code{"fill"} to indicate the display of the plot if \code{plot} is \code{"barplot"}.}

\item{x_axis_col}{Name of the categorical variable to reflect in the plot.}

\item{y_axis_col}{Name of the numerical variable to reflect in the plot.}

\item{condition_col}{Name of the categorical variable to group the data points.}

\item{ncols}{Numeric; number of columns for arranging panels. Defaults to 1.}

\item{facet_col}{Optional; name of the categorical variable to divide the plot into different panels. Required if \code{ncols} bigger than 1.}

\item{add_stats}{Logical; if \code{TRUE}, performs a statistical analysis based on the available parameters. Defaults to \code{FALSE}.}

\item{color_palette}{Optional; a vector of color codes to customize plot appearance.}

\item{targeted_arg}{Optional; a vector defining key feature clusters to highlight or label.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"}, \code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of the legend.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the requested distribution plot. If \code{save_format} is provided, the plot will also be saved to the specified location.
}
\description{
This function generates a visualization of codon-anticodon usage or amino acid demand-supply distributions across conditions.
The input data (\code{data}) is expected to be in long format and contain a minimum information regarding the features, the conditions they belong to and their usage counts.
Based on the selected \code{plot} parameter the user can select the most convenient layout to display the data.
It is crucial to ensure consistency between the name of the columns in \code{data} and the parameters \code{x_axis_col}, \code{y_axis_col} and \code{condition_col}.
}
