% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{tTE_ProportionPlot}
\alias{tTE_ProportionPlot}
\title{Proportion Plot of Codon/Anticodon Usage or Amino Acid Demand/Supply}
\usage{
tTE_ProportionPlot(
  data,
  plot = "bar",
  var_numerical,
  var_categorical,
  var_color = NULL,
  color_palette = NULL,
  save_format = NULL,
  out_name = NULL,
  out_directory = NULL,
  show_legend = "none",
  add_titles = TRUE
)
}
\arguments{
\item{data}{A long format table. This format can be obtained using \code{\link{DataToLongFormat}}.}

\item{plot}{Either \code{"bar"} (default), \code{"donut"} or \code{"radar"} to indicate the type of plot to generate.}

\item{var_numerical}{Name of the numerical variable to reflect in the plot.}

\item{var_categorical}{Name of the categorical variable to reflect in the plot.}

\item{var_color}{Optional; name of the categorical variable to group the data points when coloring. Required if \code{plot} is \code{"bar"}.}

\item{color_palette}{Optional; a vector of color codes to customize plot appearance.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"}, \code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of the legend.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the requested proportion plot. If \code{save_format} is provided, the plot will also be saved to the specified location.
}
\description{
This function generates a proportion plot to compare codon-anticodon usage or amino acid demand-supply frequencies across conditions.
The input data (\code{data}) is expected to be in long format and contain a minimum information regarding the features, the conditions they belong to and their usage counts.
Based on the selected \code{plot} parameter the user can select the most convenient layout to display the data.
It is crucial to ensure consistency between the name of the columns in \code{data} and the parameters \code{var_numerical}, \code{var_categorical} and \code{var_color}.
}
