% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanPlots.R
\name{tTE_ScoresPlot}
\alias{tTE_ScoresPlot}
\title{Violin Plot Displaying tTE Scores}
\usage{
tTE_ScoresPlot(
  data,
  conditions,
  name_sep,
  targets,
  color_palette = NULL,
  save_format = NULL,
  out_name = NULL,
  out_directory = NULL,
  show_legend = "none",
  add_titles = TRUE
)
}
\arguments{
\item{data}{A tTE results table obtained from \code{\link{Compute_tTE}}.}

\item{conditions}{A character vector specifying condition labels for samples.}

\item{name_sep}{A string delimiter used in \code{conditions} to separate the sample group labels.}

\item{targets}{A table defining key feature clusters to highlight. The first column should contain the conditions to select, and the second column the labels to use in the comparisons.}

\item{color_palette}{Optional; a vector of color codes to customize plot appearance.}

\item{save_format}{Optional; either \code{"png"} or \code{"pdf"} to specify the format to save the plot.}

\item{out_name}{Optional; name for the saved plot (if \code{save_format} specified).}

\item{out_directory}{Optional; path to the directory where the plot will be saved (if \code{save_format} specified).}

\item{show_legend}{Either \code{"none"} (default), \code{"top"}, \code{"bottom"}, \code{"right"} or \code{"left"} to specify the position of the legend in the plot.}

\item{add_titles}{Logical; if \code{TRUE}, includes titles in the plot. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object representing the tTE scores. If \code{save_format} is provided, the plot will also be saved to the specified location.
}
\description{
This function generates a violin plot to visualize the distribution of tTE scores across different condition. The tTE scores should be obtained using \code{\link{Compute_tTE}}.
The plot allows the user to easily compare the tTE distribution between different conditions, with the option to highlight specific feature clusters based on the provided \code{targets}.
}
