% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanRObject.R
\docType{class}
\name{tTEscanR_Object-class}
\alias{tTEscanR_Object-class}
\title{The tTEscanR Class}
\description{
The \strong{tTEscanR} object is dynamically updated to store \code{assays}
and \code{meta.data} at each analysis step. Ensures efficient tracking and
organization of inputs and outputs throughout the pipeline. In order to
ensure robustness throughout the pipeline, specific ids have been assigned
and should be respected by the user.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of assays.}

\item{\code{meta.data}}{A list of meta-information associated with the assays.}
}}

