% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTEscanRObject.R
\name{updateObject}
\alias{updateObject}
\title{Update Assay in a tTEscanR object}
\usage{
updateObject(
  object,
  counts = NULL,
  assay = NULL,
  meta_data = NULL,
  sample_id = NULL,
  params = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{An existing \code{tTEscanR_Object}.}

\item{counts}{A count matrix or data.frame to append or update.}

\item{assay}{Character string specifying the target assay name.}

\item{meta_data}{Optional; a data.frame of sample metadata}

\item{sample_id}{Optional; a character string naming the column in
\code{meta.data} to use as sample row names.}

\item{params}{Optional; a list of global run parameters.}

\item{overwrite}{Logical; if \code{TRUE}, overwrites existing assays or
metadata if the label coincides. Defaults to \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE}, displays information messages.
Defaults to \code{TRUE}.}
}
\value{
An updated \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
object.
}
\description{
Updates an existing \code{\link[MultiAssayExperiment]{MultiAssayExperiment}}
object by adding or replacing one or more count matrices, sample metadata, or
global parameters.
}
\examples{
data("default_tTEscanR_mRNA_data", package = "tTEscanR")
data("default_tTEscanR_tRNA_data", package = "tTEscanR")

tTEscanR_obj <- createObject(
    counts = default_tTEscanR_mRNA_data,
    assay = "mRNA",
    sample_id = "conditions"
)
tTEscanR_obj <- updateObject(
    object = tTEscanR_obj,
    counts = default_tTEscanR_tRNA_data,
    assay = "tRNA"
)
}
